+++ ResourceManager/ResManAPI/Resources/MeshResource.h
@@ -16,7 +16,7 @@ private:
 	unsigned int m_indexCount;
 
 public:
-	MeshResource(std::vector<float>& vertices, std::vector<float>& indices, const long GUID);
+	MeshResource(std::vector<float>& vertices, std::vector<unsigned int>& indices, const long GUID);
 	virtual ~MeshResource();
 
 	const sg_buffer& getVertexBuffer() const;
